+++ SasquatchMac/SasquatchMacObjC/AppCenterDelegateObjC.m
@@ -21,6 +21,10 @@
   return [MSAppCenter setEnabled:isEnabled];
 }
 
+- (void)setCountryCode:(NSString *)countryCode {
+  return [MSAppCenter setCountryCode:countryCode];
+}
+
 - (void)setCustomProperties:(MSCustomProperties *)customProperties {
   [MSAppCenter setCustomProperties:customProperties];
 }
@@ -89,15 +93,15 @@
 }
 
 - (void)trackEvent:(NSString *)eventName withProperties:(NSDictionary<NSString *, NSString *> *)properties flags:(MSFlags)flags {
-    [MSAnalytics trackEvent:eventName withProperties:properties flags:flags];
+  [MSAnalytics trackEvent:eventName withProperties:properties flags:flags];
 }
 
 - (void)trackEvent:(NSString *)eventName withTypedProperties:(MSEventProperties *)properties {
-    [MSAnalytics trackEvent:eventName withTypedProperties:properties];
+  [MSAnalytics trackEvent:eventName withTypedProperties:properties];
 }
 
 - (void)trackEvent:(NSString *)eventName withTypedProperties:(MSEventProperties *)properties flags:(MSFlags)flags {
-    [MSAnalytics trackEvent:eventName withTypedProperties:properties flags:flags];
+  [MSAnalytics trackEvent:eventName withTypedProperties:properties flags:flags];
 }
 
 - (void)trackPage:(NSString *)pageName {
@@ -113,11 +117,11 @@
 }
 
 - (void)resume {
-    [MSAnalytics resume];
+  [MSAnalytics resume];
 }
 
 - (void)pause {
-    [MSAnalytics pause];
+  [MSAnalytics pause];
 }
 
 #pragma mark - MSCrashes section.