+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # App Center SDK for iOS and macOS Change Log
 
-## Version 1.14.0 (not yet released)
+## Version 1.14.0
 
 ### AppCenter
 
@@ -9,6 +9,10 @@
 ### AppCenterAnalytics
 
 * **[Feature]** Preparation work for a future change in transmission protocol and endpoint for Analytics data on macOS. There is no impact on your current workflow when using App Center.
+
+### AppCenterPush
+
+* **[Fix]** Fix crash on invoking an optional push callback when it isn't implemented in the push delegate.
 
 ___
 
@@ -29,6 +33,7 @@ ___
 ### AppCenter
 
 * **[Fix]** Fix a possible deadlock if the SDK is started from a background thread.
+* **[Feature]** Add class method  `+ [MSAppCenter setCountryCode:]` that allows manually setting the country code on platforms where there is no carrier information available. 
 
 ___